fix make err change make mingw to make posix

+++ msys2/lua_build.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 INSTALL_PATH=$HOME/tools
-SOFT_HOME=$INSTALL_PATH/lua
+SOFT_HOME=$INSTALL_PATH/lualib
 BASH_DIR=$INSTALL_PATH/rc
 TOOLSRC_NAME=lualibrc
 TOOLSRC=$BASH_DIR/${TOOLSRC_NAME}
@@ -18,11 +18,15 @@ fi
 SOFT_FILE_NAME=lua-${SOFT_VERSION}
 SOFT_FILE_PACK=${SOFT_FILE_NAME}.tar.gz
 cd ~
-curl -o ${SOFT_FILE_PACK} -L http://www.lua.org/ftp/${SOFT_FILE_PACK}
+if [ ! -f ${SOFT_FILE_PACK} ];then
+  curl -o ${SOFT_FILE_PACK} -L http://www.lua.org/ftp/${SOFT_FILE_PACK}
+fi
 mkdir -p ${SOFT_FILE_NAME}
 tar -xzf ${SOFT_FILE_PACK} -C ${SOFT_FILE_NAME}
 cd ${SOFT_FILE_NAME}/${SOFT_FILE_NAME}
-make mingw
+make clean
+#make mingw
+make posix
 cp -R src ~/tools/lualib
 #--------------------------
 if [ ! -d "${BASH_DIR}" ]; then
